fix/activate f7ee1fe code - tmp nums mapped in DB
We're not yet sending verification codes from sgx-catapult, but all
the data is now in the database, including the temporary numbers and
their mappings.  We also have working "JID already registered" checks.

Thanks to sgx-catapult commit e856b2f this jmp-register code works
with the sgx-catapult schema, which now uses a single number value
rather than a list of size one to store the JID for a given number.