phone verification using Catapult and jmp-fwdcalls
The phone verification works by placing an API call using the
Bandwidth AP (Catapult) endpoint that initiates the phone call to the
user's fwdphone and passes along the jmp-fwdcalls URL so Bandwidth AP
will make a request to jmp-fwdcalls when the call is answered.

Thanks to jmp-fwdcalls commit 94ccea8, the call is indeed answered and
then the phone verification code is provided using text-to-speech (see
the commit for details).

Note that the verification is not yet completed, ie. the fwdphone is
not yet moved from the reg-phn_maybe-[sid] key to the _good-[sid] key.

This commit also includes a couple minor fixes to wording/indentation.