give user location-based phone number suggestions
The onboarding process has been overhauled so that users don't have to
type in an area code to start the signup process.  Instead, they're
presented with a list of phone number options based on their location
(with a variety of fallbacks in case no numbers are available there or
they're not in Canada or the US).

The user can still choose to enter an area code and pick from the
resulting list, but this is no longer the primary way to select a
number for signup purposes.

Along with the main index page changes (which include a viewport fix
so the <iframe/> looks better) and the addition of num_find.php for
the actually finding the numbers to suggest, register1 and register2
have been updated so they save/show the city the number is from.  This
helps those using the number suggestions so may as well add elsewhere.

The duplicate Signup section on the main page has been replaced with a
"Signup today!" button, both because the Signup section is now quite
long, and because we don't want to hit num_find.php twice for a user.

We do add a dependency here, on having a MaxMind account for their
GeoIP2 Precision Web Services tool.  In testing, we found that gratis
geolocation tools were not accurate enough (especially in Canada),
making it difficult to provide adequate suggestions.  The code is easy
enough to rip out (just remove the <iframe/>, fix the text).

Note that num_find.php includes a few hacks for translating MaxMind's
very good data into Catapult's less good naming conventions.  More
will probably be needed, but what we have is a good start.

Thanks to @singpolyma for the initial idea of number auto-suggestion.