fix "try again" so sip: not double-added; fixes #4
This bug was introduced in 7dea1d3 along with the rest of SIP support.
The issue is that we've always used the normalized number in the "try
again" links, which was fine when we just had E.164 numbers, but now
normalization also includes adding "sip:" in some cases.  But we can't
know whether a string starting with "sip:" has been normalized already
because a SIP address could start with "sip:" (giving a "sip:sip:..."
URI) per https://tools.ietf.org/html/rfc3261#section-19.1.1 .  So we
have to be careful to only use the string the user already gave us
(pre-normalization) in our "try again" links.