auto-register with Cheogram using XEP-0297 fwding
Thanks to 176d00d (which converted register4 to Ruby) we were able to
implement proper Cheogram integration here.  In particular, we now
automatically register each user with Cheogram immediately after the
JID verification step, assuming they will eventually complete signup.
(This assumption will be moot once jmp-register offers trial periods.)

We had to update register3 as well, since it is necessary that the
verification code be delivered from a Cheogram-accessible JID after
the signup has completed.  So we replaced the (very hacky) previous
method of making up an invalid phone number that sgx-catapult uses
when instructed to by a faked incoming SMS from Bandwidth AP with a
much simpler direct SMS to Cheogram that delivers the code to the JID.

Note that this commit changes the meaning of $sgx_url - it is now used
to define the URL for sgx-catapult's delivery receipts endpoint rather
than the Bandwidth AP "Messaging Callback" URL that sgx-catapult uses.

Cheogram supports XEP-0297 (Stanza Forwarding), which we use here to
deliver the registration request on behalf of the user who is
registering through jmp-register.