yet another fix for a9a5b83 Cheogram registration
In 843ea8d we changed the code so "Cheogram is last", but it wasn't
actually last, just after the sgx-catapult registration.  What came
after the Cheogram registration was the number purchasing.  However,
that is incorrect because Cheogram needs to be able to send an SMS
immediately after registration (it sends an "/addjid" message).  And
that's reasonable, since we claimed the sgx-catapult registration was
complete when it actually wasn't (since the number wasn't purchased).
So actually fix it for good this time; seems to work fine in staging.

What we should've done in 843ea8d was swap the moved portion of code
(the sgx-catapult registration code) with the Cheogram registration
code instead of only moving that the former.  This change fixes that.