canonical JID must be lowercase; missed in a9a5b83
When a user tried to register a JID that included uppercase characters
the non-lowercase form would be stored in the database.  Then when the
Cheogram registration was attempted, it would fail, since the JID it
sent (post-normalization, ie. lowercase) wouldn't match the one in the
database already.  So we need to ensure that we do this normalization
ourselves before storing it in the database.

Also, when the user later attempts to upgrade to a paid account, we
need a similar conversion to ensure we can match the JID stored in the
database if they happen to use a non-lowercase form.

This fix wasn't "missed in a9a5b83" per se - it was just never done at
all before, but it was something we should have thought of when adding
the other JID normalization code (resource elision and trimming).