retry when rate limit hit (easy thanks to 231a33f)
In practice, Catapult only blocks us for one second when we hit the
number search rate limit.  So wait for two seconds and then it's very
unlikely we'll hit the rate limit again.

In the future we may want to check the X-RateLimit-Reset header per
Catapult's recommendations (which provides a UTC millisecond timestamp
indicating when the rate limit expires), but this is simpler and works
just as well for now.