add XMPP servers page - spurred by xmpp.net demise
As pointed out by @keverets recently, the xmpp.net server is still not
operational (it went down two weeks ago per the discussion starting at
https://mail.jabber.org/pipermail/operators/2017-March/002956.html ).
And thus some of the links on https://jmp.chat/ were broken.  This
commit fixes those links, primarily by changing them to point to the
new suggested XMPP servers page that was created for (and included in)
this commit.

The suggested XMPP servers page includes all the servers listed on
https://gultsch.de/compliance_ranked.html that have fully-green boxes
(aside from XEP-0368, which has been deemed less important for the JMP
use case) and that support in-band registration, per the testing I did
with Gajim today, and have gratis accounts.