add voicemail support to docs, JMP SIP by default
Aside from a two-line logic change, this commit is exclusively textual
changes.  The logic change that's added here causes jmp-register to
always set a forwarding address by default, specifically the SIP
address of the JMP SIP account that's already created automatically.
This is reasonable to do now that we have voicemail support (see
https://gitlab.com/ossguy/jmp-fwdcalls/merge_requests/2 for details),
since callers will get voicemail instead of a busy signal if the
forwarding address (ie. JMP SIP account) has no logged in clients.

The bulk of the changes here are to replace text like "Voicemail...
is in the works" with text that describes the voicemail feature, and
text that described default call behaviour so that we include mention
of how voicemail works in each call receipt scenario.  These changes
are primarily to the main page (index.html) and to register4, which
is the "success" page users receive once their JMP number is acquired.

If we ever fix https://gitlab.com/ossguy/jmp-fwdcalls/issues/10 we'll
need to update jmp-register again, as the #voicemail section added
here assumes the current handling of FN and NICKNAME.