can now add non-JMP number in Catapult acct to JMP
Previously the only way to register a number with JMP through
jmp-register was to buy a new number from Catapult and then assign it
to JMP.  This makes it difficult for a user to attach a number that is
already part of the Catapult account to JMP (the user/admin would need
to execute several manual steps instead).

Now, if the number was not able to be purchased from Catapult, we will
see if it exists in our account already instead of returning an error
immediately.  If it exists and is not used by JMP yet, then we let the
user register it.

There are a few use cases for this, one being to allow users to "buy"
numbers from an internal inventory rather than from Catapult's list of
available numbers.  More on that in future commits.

This commit is almost purely additive, aside from a comment change and
control flow swap.