offer numbers from internal inventory when desired
Prior to this commit, jmp-register only ever offered or suggested
numbers to prospective users that were currently available in the main
(unpurchased) inventory provided by Catapult.  No caching or other
non-realtime handling of this inventory was done.

Here we add an option for the jmp-register administrator to include a
specific set of numbers (that they own already in their Catapult
account) to prospective JMP users, in addition to the numbers that are
already offered from Catapult's main (unpurchased) inventory.

In particular, we update both the area code search (register1) and
number suggester (num_find.php) to include numbers from the specified
list when appropriate.  The logic is as follows:

* the last number suggested by the number suggester will be replaced
  by a "fancy" number if (a) the suggester is returning two or more
  numbers in total, (b) the prospective user has at least one area
  code from $fancy_area_code_neighbours in their number list already,
  and (c) the suggester is able to find a "fancy" number from the
  administrator's list that is not used by JMP yet after checking at
  most three random numbers in the list to see if they're used by JMP

* when a user searches for numbers in area code 212, register1 will
  return up to $fancy_max numbers from the administrator's "fancy"
  number list ($fancynums[0]) that are not used by JMP yet

Note that there is still the usual race where a user may pick a number
at the start of registration that is no longer available by the time
they've verified their JID.  This is handled with an error message and
an option to start again (and select a different number), regardless
of which type of number the user attempted to acquire, "fancy" or not.

As implemented, only one set of "fancy" numbers can currently be
specified.  The eventual goal is to allow multiple sets, each with its
own distinct set of neighbouring area codes.  Each area code in the
list of neighbouring area codes ($fancy_area_code_neighbours) would
then point at the index of the list of "fancy" numbers it's adjacent
to inside of $fancynums.  Currently the code assumes that there is
only one such list ($fancynums[0]) and the values of the keys in
$fancy_area_code_neighbours are unused (responsible administrators
should set all those values to 0 for forward compatibility).

register1 currently assumes that the "fancy" numbers are in area code
212 (per the bullet point above).  For now, additional area codes
would need to be added manually to the logic in register1 (in addition
to being added to fancynum-jmp.php per the README file).

This commit was made possible by 9f39ccd and e8dc5c0.  Thanks to
9f39ccd, it is a purely additive commit.  e8dc5c0 allowed users to
register numbers already in the administrator's Catapult account, but
until this commit such numbers were not shown/selectable in the UI and
the user would have to edit the query string (ie. the number parameter
of the register2 URL) in order to register such a number.