ask user to add support JID if first verify fails
Some servers (and some clients, ie. using Gajim's "Ignore events from
contacts not in the roster" setting, which is off by default) will not
receive messages from JIDs that are not in the user's roster.  As a
result, the verification message might not be received by the user who
is attempting to signup for JMP.

So we add some text to the page where we ask for the verification code
that tells the user to first add the support JID to their roster
before retrying.  This should improve the registration process for a
select number of users.

See #12 for more details on this issue.  We won't close it yet because
there should probably still be some text to explain to the user that
their experience will be severely limited if their client or server is
configured this way, since they won't be able to receive text messages
or voicemail from phone numbers that are not in their roster.