simplify front page, make apps easier to find/get
This change primarily simplifies the main page, so that a first-time
visitor doesn't have to wade through a wall of text before getting to
the Signup section, and so that the Signup section is much easier to
follow (now a simple 3-step process!).  Links to the apps and servers
that users are likely to use with JMP are now very prominent (and have
their own icons, which are also included in this commit).

Multiple sections of the FAQ have been updated, mostly to include some
of the text that originally existed above the FAQ, but simultaneously
with their own simplifications so that the overall size of the FAQ
section did not grow (much).  The most substantial overhaul was to Q2
(the #jabber section), with Q1 (#features), Q9 (#calling), and Q10
(#voicemail) also receiving significant updates.

In particular, we are now recommending Movim as the way to update the
voicemail settings (in the #voicemail section), since it now supports
ad-hoc commands, rather than Gajim, which doesn't easily run on macOS.

Perhaps the biggest high-level change is that iOS apps are now listed
alongside Android apps whenever they are mentioned.  Previously it was
difficult for an iOS user to find the appropriate app to use.  We now
recommend ChatSecure and Linphone for XMPP and SIP, respectively, as
we have found that they work reasonably well in testing.

This commit does not include any "functional" changes to the signup
flow - all of the register* pages work exactly as they did before, but
with added/simplified text in some cases (on register2 and register4).