add new (hidden) one-time payment option, upgrade0
The new upgrade0 page provides users with a one-time payment option.
It is a slightly modified version of upgrade2 (which handles the
subscription payment options) and directs users to upgrade3 just like
upgrade2 does.  The text has been updated to mention that the upgrade3
text might not be fully accurate for the one-time payment being done
here, and to generally say that this is a one-time payment being made,
not a subscription.  A new PayPal tags setting has been added,
$paypal_tags_one_year, similar to the exists settings for the
subscription options, which holds the tags needed to display the
appropriate PayPal button for the one-time (and one year) payment.

Note that the PayPal staging URL for the form was not used here, since
a staging button for one-time payment has not been setup yet.  This
should be fixed eventually, but is not a huge deal for a hidden page.

This new page is not intended to be linked from the rest of the JMP
site at least until upgrade3 has been updated to reflect that a
one-time payment may have been used.  It is only to be used for
payments where the user has already been made aware that this is a
custom payment option that is not normally provided.