dismail.de now preferred on suggested servers page
Recently jabjab.de started silently ignoring messages from contacts
not in its users' rosters so it is no longer suitable as a preferred
or even suggested server.  As a result, we are replacing it with
dismail.de, which does support such messages and intends to continue
supporting them for the foreseeable future.

While we're at it, also rearrange the page so that the servers are
listed in their order on https://gultsch.de/compliance_ranked.html as
that makes the order a bit less arbitrary.  Additionally, remove
jabberzac.org since it apparently no longer supports MAM.  So now our
suggested list is down to 8 servers, from 10 before this change.