main page uses dismail.de instead of jabjab.de now
As a follow up to 7faa6c2, the main page now recommends dismail.de
just like the suggested servers page.  We waited to commit this one
because we didn't have a logo ready when 7faa6c2 was committed, but
thanks to the dismail.de admin, who graciously created such a logo for
us to use, we can now include it in this commit.