for iOS recommend Tigase Messenger, not ChatSecure
ChatSecure has been non-optimal for a while now, primarily because it
does not support receiving messages from contacts not in the user's
roster (see https://github.com/ChatSecure/ChatSecure-iOS/issues/844
for details).  We recently found that Tigase Messenger does support
this, and generally works at least as well as ChatSecure in other
areas (especially in that it is free and open source software) so it
makes sense to switch our recommendation.

Tigase Messenger's support for Push works best with Tigase's own
server extension, which one can use by picking the tigase.im server
when creating an account.  This is not ideal (especially since Push
without the extension seems to work not quite as well as ChatSecure),
but it's more important that non-roster messages get through for now.
In the future we can modify the signup page to recommend a tigase.im
JID when using Tigase Messenger, if they haven't fixed Push by then.