port requests now via form, linked from main page
This change adds a porting form that users can now use instead of
manually delivering the porting information through the JMP support
avenues.  The user needn't interact with JMP support at all now, as
the port will happen "automatically" once they submit the porting
information.

In order to make this happen, the porting form delivers the parameters
from the user to the JMP administrator via email, using a hack similar
to the one in 11a0c5d where $notify_receiver_email is abused for this
purpose (we should have separate variables for each type of email
address, but for convenience we are just using one for now).  The form
code also constructs a URL that includes these user-entered form
parameters so that the URL can be visited by the JMP administrator to
view the completed Letter of Authorization, with all of the requisite
values populated.  The URL's base address is specified in the new
$porting_form_url variable in the jmp-register settings file - the
contents of the page/code at that URL are beyond the scope of
jmp-register, since the form is carrier-specific.  The administrator
will need to provide their own porting page that populates its
values based on the GET parameters sent by porting3 if they wish to
take advantage of the porting form.  Normally this page would then be
submitted to the administrator's carrier, either via API, or through a
support ticket or similar.

The #existing section of the main jmp-register page now links directly
to the new porting wizard, as does the "bring your own number" link in
the Signup section.

It is up to the administrator of the jmp-register instance to submit
the Letter of Authorization once they receive the email sent by the
porting3 page, and then follow up with the JMP user once a FOC date
has been established.  Additional manual steps are required in order
to switch the user's JMP number once the port completes - such steps
are beyond the scope of jmp-register for now.