bump beta exit one(?) more time; last was c20a09e
Getting an account at a suitable payment processor has proven to be
more difficult than anticipated - we haven't been able to start the
core development work because we don't have an account yet (and thus
we don't know which API we will ultimately be using, as we can't say
for sure which of our options will pan out).  Things are looking up
with one processor, though, so we should be out of the woods soon.