set start day for a number's usage stats at signup
In order for the "Show Monthly Usage" command of jmp-fwdcalls to work
correctly, it needs to know when a given number was last registered so
that it doesn't show stats from possible previous registrations when
it returns the stats to the current user of a given JMP number.  As a
result, we need to set the date that the number was registered as part
of the registration process, which is what this change does (setting
the value of usage_start_day-[JMP_number] to today's date - ideally
this would be a TAI date, but for now we're just using UTC because it
is close enough for our purposes, and adding ./tai as a dependency (as
we did for jmp-fwdcalls and sgx-catapult) seems like a bit much; note
that the server's timezone must be UTC as we still need a fix here).

For further details on how usage_start_day-[JMP_number] is used, see
https://gitlab.com/ossguy/jmp-fwdcalls/commit/5b1481b8876d8a3c78ba1024adb2e5c12be0cf06 .

With this change to jmp-register, all subsequent registrations will
have the correct information for the "Show Monthly Usage" command, but
if some users registered before this change was deployed, then their
usage_start_day-[JMP_number] values MUST be set manually.