"trial accounts need wait/referral" part 1, of #18
This is the first part of implementing the "trial accounts need
wait/referral" sub-bullet in ticket #18.  In particular, we are only
sending the signup requests to a JID for now - there is no logic yet
to block the trial signups until they are approved, so they go through
just as before.  But now at least there's a JID that sees when they're
happening.

In the future these "signup requested" messages will be something that
an administrator can act on, either by approving, denying, or ignoring
the request.  For now they are just informational.

This commit adds a new setting, which lets the administrator specify
where they want these "signup requested" messages to go, hence the
update to README.creole .

Note that registere/index.php is a very lightly modified version of
register3/index.php - the only changes are related to sending the new
"signup requested" messages, and some additional "TODO NOW" items that
explain how we will be implementing the rest of the signup changes.

In particular, these "TODO NOW" comments describe the rest of how we
plan to implement the "support for new payment methods" bullet in #18.

The signup process now has a "Processing registration..." step that is
used to send this message.  We will eventually change this page into a
list of options that the user can select (to either pay, use a
referral code, or wait for a trial account to be manually approved).
When that's done the "Processing registration..." step will be moved
one level down, but for now we'll leave it at the "top level" for ease
of coding.