major 2nd step of painful signup decoupling - #18
The first major step was in 4a26799, which laid the groundwork for
this.  Thanks to a448170 we didn't need to include any renames here so
it's a fairly explanatory commit.

At a high level, we are trying to decouple the various bits of the
signup process, most of which have been housed in register4 until now.
We are doing this so that it is easier for a user to enter the process
at any given step, which makes it easier to add things like payment
options and the "trial accounts need wait/referral" feature mentioned
in #18.

With this commit we are finally at the point where we at least have
the steps for the base case in their final location.  In particular,
the options1 step is intended to complete the JID verification (which
it does now) and then give users options for how they want to pay/etc.
for their JMP account (which it will do later).  The brand new
wait_for_approval1 step will give users the option to wait for an
admin to approve their JMP trial signup (which is the closest
approximation to the current behaviour, where it's auto-approved), but
for now just continues the auto-approving that was done already.
However, we can much more easily add the "wait" feature, since the
user can just keep reloading this page now, as it only does one thing.

The major changes to make this all happen are:

1. Moving a bunch of stuff from register4 into options1.  Of special
note is the JID verification, which now happens in options1 so we know
at that point whether we can "guarantee" that the person with a given
sid (session ID) owns a given JID.  This makes it easier to reason
about the next steps.

2. Moving some stuff from options1 into wait_for_approval1 (the reason
we needed to create it to begin with).  This is really just the signup
notification code, which lets a custom JID know about new signups that
may need approval.  For now it's notify-only, but eventually will
block a trial signup until the approval is granted.

3. Add extra verification steps to register4.  Now that we have the
JID verification done in another step, we need a different mechanism
to tell us that the registration can go through.  And that mechanism
is the payment-plan_as_of_{YYYYMM}-{jid} key, which we check to
confirm that the user is authorized to receive a JMP trial account.
This is just auto-set by wait_for_approval1 right now, but will
eventually be set only when an admin approves the JMP trial account.

The above required a bunch of rewriting Ruby in PHP, but it seems to
have all worked out fine, as it was fairly straight-forward.  There
are many other changes to make (and especially cases not yet properly
considered by the code as-written), but these will become easier to
write as we start building the other options into what we have now.