fix JMP number registration - avoid getting wedged
With the previous placement of the credential setting step there would
often arise a situation where the JMP number wasn't available anymore
but the credentials were setup already so the user couldn't try again
with a different number (since the credentials were there even though
the JMP number was not), causing their JID to be "wedged" in the
system - i.e. in a place where it couldn't be registered anymore, but
couldn't be used with the system either (since the number hadn't been
bought successfully) unless an administrator fixed it manually.

Now we don't setup any credentials until the number has been bought or
otherwise acquired.  I suspect the original reason for setting up
credentials first was so we didn't buy a number unnecessarily (in case
the credential setup failed).  But we've noticed in practice that this
is not much of a concern so it's far better to prevent the wedge case.

We've been running a variant of this in production for a while, but
unfortunately it only moves the catapult_jid-* setting, so it's maybe
a bit worse (as the wedge is "partial" - i.e. only the credentials and
not the catapult_jid-* key).  That was a temporary attempt at solving
the issue when we discovered it, but not very successful.  Now we have
a better-tested and more thoroughly analyzed fix (in this commit).