add reg-pending_approval_session-* keys for admins
Add new keys to assist in the approval process for new trial signups.
A modulo of these will be used to make it quicker for the admin to
pick which JID they want to approve (based on which session that JID
has been verified in).  More on that in an upcoming commit.

Note that the TTL is being doubled here just to make sure that we are
not expiring this new key before the session key expires.  This should
only happen in some extremely rare race condition if at all, but may
as well be safe.