ask user to wait 30 seconds instead of 5 for loads
While in most cases the wait time to the next screen will be less than
5 seconds, sometimes the next page takes a few seconds to load so on
top of the 3 seconds that the refresh waits it feels too long for the
user (as we have seen empirically with a bunch of users clicking too
prematurely).  So ask them to wait a bit longer before clicking so
they don't, for example, fail to see the SIP account info because they
clicked the link at the same time it was starting to load.

The arbitrary "5 seconds" was first added in 4a26799 and then copied
in 96b6853 - this commit fixes both of those.