move signup rate limiting so can still show errors
Since nothing signup-related is written until later, and since we have
a better error message as of 96b6853 (for cError), it is better that
we move the signup rate limiting down, so the user can see this better
error message instead of getting unnecessarily rate-limited (since no
signup is occurring if this rate limit error is shown).

Now the signup rate limiting is done immediately before the first part
of the actual signup, as it should be.