soft launch of cryptocurrency payments; not on idx
We are adding the bare minimum required to support cryptocurrency
payments in an automated way in this commit.  In particular, we have a
way of accepting them (for 1-year or 3-year terms), and of showing the
progress of the payment itself, but if the user wants to know when
their next payment is due, they still need to contact us (just like
before).  That will be automated later - just baby steps for now.

Since this has not been widely tested yet, we are going to hold off on
advertising it on the main (index) page.  We have a bunch more copy to
write there anyway before we launch it, so save that until later.

We got the "pay with Bitcoin" icon added in this commit from
https://lukaszadam.com/illustrations - the page says that the icons
are "Completely free for commercial projects - no attribution
required."  But we did attribute Lukasz Adam in the alt text and
filename anyway, since it seemed like the right thing to do.  The
specific icon set we used was this one:

  https://lukaszadam.com/assets/downloads/badge1.zip

Note that this commit does depend on some external code that is not
included here.  In particular, we rely on the following to render the
payment page (defined by $electrum_url_prefix):

  https://github.com/spesmilo/electrum-merchant

The above does not work as-is with the latest Electrum - some changes
are required.  We will be adding these in a later commit.

Another noteworthy change here is the addition of the upgrade4 page.
This is used to create the payment request so that we don't have to do
this on the upgrade2 page (which would cause a bunch of unnecessary
requests to be created).  This page is a bit odd in that it's after
upgrade3, but upgrade3 is only for PayPal and upgrade4 is only for
cryptocurrencies (the two don't talk to each other).  It seemed like a
reasonable expedient name to pick, though, so we're sticking with it.