fix empty sid for register5; regression in 96b6853
A user reported to me that they could not set a forwarding number
using register5 anymore.  It turned out this was because of a
regression in 96b6853, namely that @cleanSid had been deleted.  We
thought this was fine because it was no longer used anywhere in
reg_cheo.rb but failed to remember that it was still used in the
success.rb view.  So this restores @cleanSid, moving it to a more
appropriate location in the process (in hopes that it won't be
accidentally removed again, due to its proximity to other variables
that are of a similar nature).