Set up notification
Since each request has a unique address, any trancation coming to that
address is for that request.  Note that this means the URL could be hit
for a transaction that is too small to actually fill the request, so we
should probably check the actual status in the callback script.