New pages and styles
This is basically a rewrite of the homepage, splitting it into homepage, FAQ
page, and new first-steps of the register process.

Content and styles and markup have been reworked to the point of rewrite on
those pages, and some existing pages have been updated to integrate the basics
of the new styles, but not everything is 100% consistent yet (not that it was
before).

The hamburger nav appears on all new pages and some existing pages, it should
maybe go on every existing page as the styles get rolled over them all, but
there maybe be some exceptions.

Credits for new icons, etc, are in the new /credits/ page.