electrum_notify.php no longer tied to JID for phone number
For one-offs as we use it now with electrum payment requests, this was ok, but
once we move to deposit addresses it's a disaster since JID and/or phone number
may change over time.  So the notify script should always receive the stable
customer_id.  This also means BTC payers will get a customer_id set which will
be useful generally for adding them to billing system.

This change requires that jmp-settings.php have two things added:

$brantree_config = [
	'environment' => 'production',
	'merchantId' => '',
	'publicKey' => '',
	'privateKey' => ''
];

$hmac_key = 'generate a random string and store it here';