match Cheogram registration timeout to production
Since it sometimes takes a while for Cheogram to reply to registration
requests, we've been using a longer timeout value in production for a
while now, but hadn't committed it.  In particular, here is the
timeline of changes:

2017-12-02T23:03:27Z - updated from 5 to 10 seconds
2017-12-03T21:52:06Z - updated from 10 to 30 seconds

Since we've been using this new value for a little over 3 years now,
probably best to put it in the repository as it seems to work fine.