turn off auto-approvals, add Bitcoin auto-accepts
We turned off automatic trial account approvals in stages between
2020-12-06 and 2020-12-08 in production.  This commit reflects those
changes, specifically the wait_for_approval1/index.php deletions,
which cause the signup to not complete until a separate process has
put the payment plan key in place.

We also updated the text on the last page of signup over time, first
to reflect how long a user would have to wait for trial approval (at
2020-12-06T22:08:10Z - "We reply within 9 hours"), and then once we
added Bitcoin auto-accept logic (at 2020-12-17T05:13:57Z - "Currently
registration requires either manual approval, or payment via Bitcoin")
- the main part of which was in fd1dfb2 (the part made to upgrade4,
which is what we now link to in the new text in this commit) but also
in b8590e4 (the part that actually sets the payment plan, which is
required in order for the account to be auto-enabled).

While there are still some fixes to b8590e4 that have yet to be added,
this commit brings us in line with the user-facing production changes
per above - the rest of the changes are mainly for debugging purposes.