update Bitcoin auto-accept price & initial length
We have new production values for the price and initial length given
for people who use the Bitcoin auto-accept option to get a new JMP
account.  In particular, here are the timestamps of the updates:

2020-12-30T01:22:12Z - number of months increased from 4 to 5
2021-02-21T23:52:25Z - price decreased from 0.599 mBTC to 0.25 mBTC

Note that the last change was done as part of 485eab8, but was not
committed at that time.  With this commit we are now in sync again.