update signup success page since can be non-trial
The success page was updated in production to reflect the Bitcoin
auto-accept feature at a few different times, per some of the changes
made in d63b9c4.  In particular, the main changes were made on
2020-12-21 to reflect that the user might not have a trial, and that
the initial payment term is 4 months, and then the 4-month length was
increased to 5 months at 2021-02-21T23:33:31Z to reflect the changes
described in 6344bce (a bit late, so it was out of sync for a while).

Now the text reflects what we are using in production per the above
timestamp history.