update upgrade4 to handle the Bitcoin auto-accept
This should have been committed prior to d63b9c4, as that change
depends on this one to work.  In any case, this change lets us pass
through some extra parameters to our Electrum system, so that it can
show the user a link if the user is using the Bitcoin auto-accept
method to start their JMP account (as opposed to waiting for trial
account approval).  Once the Bitcoin payment is completed, our
Electrum system will construct a link based on the number and session
ID passed through here, so the user can click on that to complete
their registration (rather than having to start from the beginning).

The other branch of the control flow (what we had here before) is
intended for people who are paying via Bitcoin that already have a JMP
account (i.e. they have a trial account or they are adding funds to an
existing paid account).