correctness updates to b8590e4, plus notifications
This updates the initial electrum_notify.php (added in b8590e4) to fix
a few issues, such as:

* can't handle non-approved user (needed for auto-accept per d63b9c4)
* no '?>' at the end
* settings file inclusion path not quite correct
* catapult_jid key not quite correct
* cheo_jid not created correctly
* only sets current month payment plan (might fail since next mo soon)

In addition to fixing all of the above, this change also sends a
notification to the $notify_receiver_email when a new payment event
occurs - there are still some issues with this (triggered a bit too
often, possibly because Electrum requires us to confirm we received
the even in some way), but it at least lets us know most of the time
when there's a new payment received.

The above changes were all made between 2020-12-16 and 2020-12-20,
which was just after b8590e4 was committed.  The final change was made
on 2020-12-28, to remove some of the extra notifications we had added,
since we determined they were redundant, so they're not included here.