Wire up BTC flows to use jmp-pay
Change the electrum endpoint to jmp-pay and set things up for acconut balances.
This means asking users to choose an account balance currency before they send
BTC.  Electrum payment request only needs to be used for the initial
funding (since we have minimums there anyway) and we set a pending plan key in
redis in that case to be inserted to the plan_log table by jmp-pay after payment
confirms.

Users who have ever paid with BTC on the new system will just see the list of
their deposit addresses.