140369b patch so can use Braintree gem in CGI mode
In particular, without this patch we get the following in our Apache
error logs when using CGI:

"AH02429: Response header name 'I, [2021-03-30T17' contains invalid
characters, aborting request"

Note that the header name will vary depending on the time Braintree is
trying to output its log message to stdout (for example, the first
"header name" we saw was 'I, [2021-03-28T03').  But in short, with CGI
its log output (which by default goes to stdout) is getting prepended
in front of the actual output that we want Apache to deliver.  So we
need to turn that off, which this patch does.