Jabber SMS Transport SEO
Lots of people are looking for the thing that we are, but we don't rank highly
for it because we avoid the technical terms people are searching for.  Adding a
small FAQ entry should not be a big obstacle to less technical users and may
boost our ranking a bit for this.